+++ src/pages/home/components/form-user/form-user.tsx
@@ -7,10 +7,9 @@ import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { Cookies } from "react-cookie";
 import { Link } from "react-router-dom";
-
 import { AxiosError } from "axios";
 import { useState } from "react";
-import styles from "../../styles.module.css";
+import styles from "../../home.module.css";
 import { api } from "../../../../api";
 
 interface FormProps {
